Refactor literals
