+++ reflex/components/datadisplay/dataeditor.py
@@ -6,7 +6,7 @@ from typing import Any, Callable, Dict, List, Optional, Union
 
 from reflex.base import Base
 from reflex.components.component import Component, NoSSRComponent
-from reflex.components.literals import LiteralRowMarker
+from reflex.components.literals.data_editor import LiteralRowMarker
 from reflex.utils import console, format, imports, types
 from reflex.utils.serializers import serializer
 from reflex.vars import ImportVar, Var, get_unique_variable_name
