+++ reflex/components/datadisplay/divider.py
@@ -1,6 +1,6 @@
 """A line to divide parts of the layout."""
 
-from reflex.components.graphing.recharts.recharts import LiteralLayout
+from reflex.components.literals.base import LiteralOrientation
 from reflex.components.libs.chakra import ChakraComponent, LiteralDividerVariant
 from reflex.vars import Var
 
@@ -11,7 +11,7 @@ class Divider(ChakraComponent):
     tag = "Divider"
 
     # Pass the orientation prop and set it to either horizontal or vertical. If the vertical orientation is used, make sure that the parent element is assigned a height.
-    orientation: Var[LiteralLayout]
+    orientation: Var[LiteralOrientation]
 
     # Variant of the divider ("solid" | "dashed")
     variant: Var[LiteralDividerVariant]
