+++ reflex/components/disclosure/tabs.py
@@ -3,8 +3,8 @@
 from typing import List, Optional, Tuple
 
 from reflex.components.component import Component
-from reflex.components.libs.chakra import (
-    ChakraComponent,
+from reflex.components.libs.chakra import ChakraComponent
+from reflex.components.literals.chakra import (
     LiteralColorScheme,
     LiteralTabsVariant,
     LiteralTagAlign,
