+++ reflex/components/feedback/alert.py
@@ -1,8 +1,8 @@
 """Alert components."""
 
 from reflex.components.component import Component
-from reflex.components.libs.chakra import (
-    ChakraComponent,
+from reflex.components.libs.chakra import ChakraComponent
+from reflex.components.literals.chakra import (
     LiteralAlertVariant,
     LiteralStatus,
 )
