+++ reflex/components/feedback/spinner.py
@@ -1,6 +1,7 @@
 """Container to stack elements with spacing."""
 
-from reflex.components.libs.chakra import ChakraComponent, LiteralSpinnerSize
+from reflex.components.libs.chakra import ChakraComponent
+from reflex.components.literals.chakra import LiteralSpinnerSize
 from reflex.vars import Var
