+++ reflex/components/forms/checkbox.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 
 from typing import Any, Union
 
-from reflex.components.libs.chakra import (
-    ChakraComponent,
+from reflex.components.libs.chakra import ChakraComponent
+from reflex.components.literals.chakra import (
     LiteralColorScheme,
     LiteralTagSize,
 )
