+++ reflex/components/forms/input.py
@@ -4,8 +4,8 @@ from typing import Any, Dict
 
 from reflex.components.component import Component
 from reflex.components.forms.debounce import DebounceInput
-from reflex.components.libs.chakra import (
-    ChakraComponent,
+from reflex.components.libs.chakra import ChakraComponent
+from reflex.components.literals.chakra import (
     LiteralButtonSize,
     LiteralInputVariant,
 )
