+++ reflex/components/forms/rangeslider.py
@@ -4,7 +4,8 @@ from __future__ import annotations
 from typing import Any, List, Optional, Union
 
 from reflex.components.component import Component
-from reflex.components.libs.chakra import ChakraComponent, LiteralChakraDirection
+from reflex.components.libs.chakra import ChakraComponent
+from reflex.components.literals.chakra import LiteralChakraDirection
 from reflex.constants import EventTriggers
 from reflex.utils import format
 from reflex.vars import Var
