+++ reflex/components/forms/select.py
@@ -4,7 +4,8 @@ from typing import Any, Dict, List, Union
 
 from reflex.components.component import Component
 from reflex.components.layout.foreach import Foreach
-from reflex.components.libs.chakra import ChakraComponent, LiteralInputVariant
+from reflex.components.libs.chakra import ChakraComponent
+from reflex.components.literals.chakra import LiteralInputVariant
 from reflex.components.typography.text import Text
 from reflex.constants import EventTriggers
 from reflex.utils.types import _issubclass
