+++ reflex/components/forms/switch.py
@@ -3,7 +3,8 @@ from __future__ import annotations
 
 from typing import Any, Union
 
-from reflex.components.libs.chakra import ChakraComponent, LiteralColorScheme
+from reflex.components.libs.chakra import ChakraComponent
+from reflex.components.literals.chakra import LiteralColorScheme
 from reflex.constants import EventTriggers
 from reflex.vars import Var
