+++ reflex/components/forms/textarea.py
@@ -5,7 +5,8 @@ from typing import Any, Union
 
 from reflex.components.component import Component
 from reflex.components.forms.debounce import DebounceInput
-from reflex.components.libs.chakra import ChakraComponent, LiteralInputVariant
+from reflex.components.libs.chakra import ChakraComponent
+from reflex.components.literals.chakra import LiteralInputVariant
 from reflex.constants import EventTriggers
 from reflex.vars import Var
