+++ reflex/components/graphing/recharts/polar.py
@@ -6,12 +6,12 @@ from typing import Any, Dict, List, Union
 from reflex.constants import EventTriggers
 from reflex.vars import Var
 
-from .recharts import (
+from .recharts import Recharts
+from reflex.components.literals.recharts import (
     LiteralAnimationEasing,
     LiteralGridType,
     LiteralPolarRadiusType,
     LiteralScale,
-    Recharts,
 )
