+++ reflex/components/layout/card.py
@@ -3,8 +3,8 @@
 from typing import Optional
 
 from reflex.components.component import Component
-from reflex.components.libs.chakra import (
-    ChakraComponent,
+from reflex.components.libs.chakra import ChakraComponent
+from reflex.components.literals.chakra import (
     LiteralCardVariant,
     LiteralColorScheme,
     LiteralTagSize,
