--- reflex/components/literals.py
@@ -1,5 +1,0 @@
-"""Literal custom type used by Reflex."""
-
-from typing import Literal
-
-LiteralRowMarker = Literal["none", "number", "checkbox", "both", "clickable-number"]
