+++ reflex/components/overlay/alertdialog.py
@@ -4,7 +4,8 @@ from __future__ import annotations
 from typing import Any, Union
 
 from reflex.components.component import Component
-from reflex.components.libs.chakra import ChakraComponent, LiteralAlertDialogSize
+from reflex.components.libs.chakra import ChakraComponent
+from reflex.components.literals.chakra import LiteralAlertDialogSize
 from reflex.components.media.icon import Icon
 from reflex.vars import Var
