+++ reflex/components/overlay/menu.py
@@ -5,8 +5,8 @@ from typing import Any, List, Optional, Union
 
 from reflex.components.component import Component
 from reflex.components.forms.button import Button
-from reflex.components.libs.chakra import (
-    ChakraComponent,
+from reflex.components.libs.chakra import ChakraComponent
+from reflex.components.literals.chakra import (
     LiteralChakraDirection,
     LiteralMenuOption,
     LiteralMenuStrategy,
