+++ reflex/components/overlay/tooltip.py
@@ -3,7 +3,8 @@ from __future__ import annotations
 
 from typing import Any, Union
 
-from reflex.components.libs.chakra import ChakraComponent, LiteralChakraDirection
+from reflex.components.libs.chakra import ChakraComponent
+from reflex.components.literals.chakra import LiteralChakraDirection
 from reflex.vars import Var
