+++ reflex/components/radix/themes/typography.py
@@ -8,18 +8,17 @@ from typing import Literal
 
 from reflex.vars import Var
 
-from .base import (
-    CommonMarginProps,
+from .base import CommonMarginProps, RadixThemesComponent
+from reflex.components.literals.radix import (
     LiteralAccentColor,
     LiteralVariant,
-    RadixThemesComponent,
+LiteralTextSize,
+LiteralTextWeight,
+
+LiteralTextAlign,
+LiteralTextTrim,
+LiteralLinkUnderline
 )
-
-LiteralTextWeight = Literal["light", "regular", "medium", "bold"]
-LiteralTextAlign = Literal["left", "center", "right"]
-LiteralTextSize = Literal["1", "2", "3", "4", "5", "6", "7", "8", "9"]
-LiteralTextTrim = Literal["normal", "start", "end", "both"]
-
 
 class Text(CommonMarginProps, RadixThemesComponent):
     """A foundational text primitive based on the <span> element."""
@@ -98,8 +97,6 @@ class Kbd(CommonMarginProps, RadixThemesComponent):
     # Text size: "1" - "9"
     size: Var[LiteralTextSize]
 
-
-LiteralLinkUnderline = Literal["auto", "hover", "always"]
 
 
 class Link(CommonMarginProps, RadixThemesComponent):
